+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.0] - 2023-07-20
+### Added
+
+- Added better file management systems. Now `spelunker` will create directories for data processing and downloads.
+- You can specify a directory for `spelunker` with the `dir` parameter. Specifying a directory is now optional.
+- Added a `save` parameter for `spelunker.load` to toggle saving the `fg_array`, `fg_flux`, and `fg_time` arrays as a `.npy` file.
+
 ## [0.5.5] - 2023-07-20
 
 - Changes to setup.py. Now `pip install spelunker` should run more properly.
@@ -17,4 +24,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-- spelunker
+- `spelunker` is added to PyPI!